% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chaos.R
\name{bayes_chaos}
\alias{bayes_chaos}
\title{Bayesian Sparse Polynomial Chaos Expansion}
\usage{
bayes_chaos(
  X,
  y,
  max_degree_init = 2,
  max_order_init = 1,
  prior = c(1, 1, 1),
  lambda = 1,
  rho = 0,
  regularize = TRUE,
  max_degree = 12,
  max_basis = 1e+05,
  verbose = TRUE
)
}
\arguments{
\item{X}{A dataframe or matrix of predictors scaled to be between 0 and 1}

\item{y}{a reponse vector}

\item{max_degree_init}{The maximum polynomial degree (for initial model)}

\item{max_order_init}{The maximum order of interaction (for initial model)}

\item{prior}{A vector with prior values: n0 (strength of prior for coefficients), v0 (degrees of freedom for IG prior for sigma2), s0 (scale for IG prior for sigma2).}

\item{lambda}{A parameter fed to the KIC calculations. Larger values will lead to sparser models.}

\item{rho}{Basis functions are only kept if their square partial correlation coefficient is bigger than rho.}

\item{regularize}{Logical. Should LASSO be used to pre-sparsify the library of basis functions?}

\item{max_degree}{The maximum degree allowed (bounds the computation time).}

\item{max_basis}{The maximum number of candidate basis functions to consider (bounds the computation time)}

\item{verbose}{Logical. Should progress information be printed?}
}
\description{
The emulation approach of Shao et al. (2017), with some minor modifications (see README for details).
}
\details{
Implements a modified version of the Bayesian sparse PCE method described by Shao et al. (2017).
}
\examples{
X <- lhs::maximinLHS(100, 2)
f <- function(x) 10.391*((x[1]-0.4)*(x[2]-0.6) + 0.36)
y <- apply(X, 1, f) + rnorm(100, 0, 0.1)
fit <- bayes_chaos(X, y)
}
\references{
Shao, Q., Younes, A., Fahs, M., & Mara, T. A. (2017). Bayesian sparse polynomial chaos expansion for global sensitivity analysis. Computer Methods in Applied Mechanics and Engineering, 318, 474-496.
}
