% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_chaos.R
\name{plot.adaptive_khaos}
\alias{plot.adaptive_khaos}
\title{Plot Method for class adaptive_khaos}
\usage{
\method{plot}{adaptive_khaos}(x, ...)
}
\arguments{
\item{x}{An object returned by the \code{adaptive_khaos()} function.}

\item{...}{Additional arguments for plotting}
}
\description{
See \code{adaptive_khaos()} for details.
}
\details{
Plot function for adaptive_khaos object.
}
\examples{
X <- lhs::maximinLHS(100, 2)
f <- function(x) 10.391*((x[1]-0.4)*(x[2]-0.6) + 0.36)
y <- apply(X, 1, f) + stats::rnorm(100, 0, 0.1)
fit <- adaptive_khaos(X, y)
plot(fit)

}
\references{
Francom, Devin, and Bruno Sansó. "BASS: An R package for fitting and performing sensitivity analysis of Bayesian adaptive spline surfaces." Journal of Statistical Software 94.LA-UR-20-23587 (2020).
}
