% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_khaos2.R
\name{plot.sparse_khaos2}
\alias{plot.sparse_khaos2}
\title{Plot Method for class sparse_khaos}
\usage{
\method{plot}{sparse_khaos2}(x, ...)
}
\arguments{
\item{x}{An object returned by the \code{sparse_khaos()} function.}

\item{...}{additional arguments passed to \code{plot}}
}
\description{
See \code{sparse_khaos()} for details.
}
\details{
Plot function for sparse_khaos object.
}
\examples{
X <- lhs::maximinLHS(100, 2)
f <- function(x) 10.391*((x[1]-0.4)*(x[2]-0.6) + 0.36)
y <- apply(X, 1, f) + stats::rnorm(100, 0, 0.1)
fit <- sparse_khaos(X, y)
plot(fit)
}
\references{
Shao, Q., Younes, A., Fahs, M., & Mara, T. A. (2017). Bayesian sparse polynomial chaos expansion for global sensitivity analysis. Computer Methods in Applied Mechanics and Engineering, 318, 474-496.
}
