% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_chaos.R
\name{print.sparse_khaos}
\alias{print.sparse_khaos}
\title{Print Method for class sparse_khaos}
\usage{
\method{print}{sparse_khaos}(x, ...)
}
\arguments{
\item{x}{An object returned by the \code{sparse_khaos()} function.}

\item{...}{additional arguments passed to or from other methods.}
}
\description{
See \code{sparse_khaos()} for details.
}
\details{
Print function for sparse_khaos object.
}
\examples{
X <- lhs::maximinLHS(100, 2)
f <- function(x) 10.391*((x[1]-0.4)*(x[2]-0.6) + 0.36)
y <- apply(X, 1, f) + stats::rnorm(100, 0, 0.1)
fit <- sparse_khaos(X, y)
print(fit)
}
\references{
Shao, Q., Younes, A., Fahs, M., & Mara, T. A. (2017). Bayesian sparse polynomial chaos expansion for global sensitivity analysis. Computer Methods in Applied Mechanics and Engineering, 318, 474-496.
}
