% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol.R
\name{sobol.sparse_khaos}
\alias{sobol.sparse_khaos}
\title{Sparse Kahos Sensitivity Analysis}
\usage{
\method{sobol}{sparse_khaos}(object, ...)
}
\arguments{
\item{object}{a fitted model output from the \code{sparse_khaos} function.}

\item{...}{additional arguments passed}
}
\value{
a list with two elements:
\item{S}{a vector of sensitivity indices with number of entries the number of parameters.}
\item{ST}{a vector of sensitivity indices with number of entries the number of parameters.}
}
\description{
Decomposes the variance of the Sparse Khaos model into variance due to main effects and total effects
}
\details{
Performs analytical Sobol' decomposition for each MCMC iteration in mcmc.use (each corresponds to a MARS model), yeilding a posterior distribution of sensitivity indices.  Can obtain Sobol' indices as a function of one functional variable.
}
\seealso{
\link{sparse_khaos} for model fitting and \link{predict.sparse_khaos} for prediction.
}
\keyword{Sobol}
\keyword{decomposition}
