% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sobol.R
\name{sobol}
\alias{sobol}
\title{Sensitivity Analysis}
\usage{
sobol(object, ...)
}
\arguments{
\item{object}{a fitted model output}
}
\description{
Decomposes the variance of the Sparse Khaos model into variance due to main effects and total effects
}
