% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sparse_chaos.R
\name{sparse_khaos}
\alias{sparse_khaos}
\title{Bayesian Sparse Polynomial Chaos Expansion}
\usage{
sparse_khaos(
  X,
  y,
  degree = c(2, 2, 16),
  order = c(1, 1, 4),
  prior = c(1, 1, 1),
  lambda = 1,
  rho = 0,
  regularize = TRUE,
  max_basis = 3e+05,
  verbose = TRUE
)
}
\arguments{
\item{X}{A dataframe or matrix of predictors scaled to be between 0 and 1}

\item{y}{a reponse vector}

\item{degree}{Maximum polynomial degree parameters. c(Initial, Increment, Maximum)}

\item{order}{Maximum order of interaction parameters. c(Initial, Increment, Maximum)}

\item{prior}{A vector with prior values: n0 (strength of prior for coefficients), v0 (degrees of freedom for IG prior for sigma2), s0 (scale for IG prior for sigma2).}

\item{lambda}{A parameter fed to the KIC calculations. Larger values will lead to sparser models.}

\item{rho}{Basis functions are only kept if their square partial correlation coefficient is bigger than rho.}

\item{regularize}{Logical. Should LASSO be used to pre-sparsify the library of basis functions?}

\item{max_basis}{The maximum number of candidate basis functions to consider (bounds the computation time)}

\item{verbose}{Logical. Should progress information be printed?}
}
\description{
The emulation approach of Shao et al. (2017), with some minor modifications (see README for details).
}
\details{
Implements a modified version of the Bayesian sparse PCE method described by Shao et al. (2017).
Degree (equivalently for order) is incremented by \code{degree[2]} if the current model contains any terms of maximal degree, and the process is started again from the top. Cannot be incremented above \code{degree[3]}.
Fitting process terminates when either (i) max_basis is exceeded or (ii) current model fit does not contain any terms with degree and order not equal to the current maximum value.
}
\examples{
X <- lhs::maximinLHS(100, 2)
f <- function(x) 10.391*((x[1]-0.4)*(x[2]-0.6) + 0.36)
y <- apply(X, 1, f) + stats::rnorm(100, 0, 0.1)
fit <- sparse_khaos(X, y)
}
\references{
Shao, Q., Younes, A., Fahs, M., & Mara, T. A. (2017). Bayesian sparse polynomial chaos expansion for global sensitivity analysis. Computer Methods in Applied Mechanics and Engineering, 318, 474-496.
}
